修复传输中偶尔中断的问题，接收界面添加ProgressBar显示传输进度

+++ p2ptransfercore/src/main/java/com/merpyzf/transfermanager/entity/ApkFile.java
@@ -11,7 +11,7 @@ public class ApkFile extends FileInfo {
 
     private Drawable aplDrawable;
 
-    public ApkFile(String name, String path, int type, long size, Drawable aplDrawable) {
+    public ApkFile(String name, String path, int type, int size, Drawable aplDrawable) {
         super(name, path, type, size, "apk");
         this.aplDrawable = aplDrawable;
     }
