+++ p2ptransfercore/src/main/java/com/merpyzf/transfermanager/entity/PicFile.java
@@ -7,7 +7,7 @@ package com.merpyzf.transfermanager.entity;
 
 public class PicFile extends FileInfo {
 
-    public PicFile(String name, String path, int type, long size) {
+    public PicFile(String name, String path, int type, int size) {
         super(name, path, type, size,"png");
         // TODO: 2018/1/9 需要考虑文件后缀名的作用，
